+++ stories/02-Design System/04-DataEntry/Input/data.ts
@@ -31,6 +31,17 @@ export const argTypes = {
             }
         }
     },
+    htmlFor: {
+        name: 'htmlFor',
+        description:
+            'Indicates the form element that the label describes. If you set the `id` prop, you should set the same `id` on the `htmlFor` prop',
+        control: {type: 'text'},
+        table: {
+            type: {
+                summary: 'string'
+            }
+        }
+    },
     type: {
         name: 'type',
         description: