+++ stories/02-Design System/04-DataEntry/InputWrapper/data.ts
@@ -21,6 +21,18 @@ export const argTypes = {
             category: 'InputWrapper'
         }
     },
+    htmlFor: {
+        name: 'htmlFor',
+        description:
+            'Indicates the form element that the label describes. If you set the `id` prop, you should set the same `id` on the `htmlFor` prop',
+        control: {type: 'text'},
+        table: {
+            type: {
+                summary: 'string'
+            },
+            category: 'InputWrapper'
+        }
+    },
     required: {
         name: 'required',
         description: 'Whether the input is mandatory',