+++ stories/02-Design System/04-DataEntry/InputWrapper/examples/index.tsx
@@ -4,6 +4,7 @@ import Border from './border';
 import Disabled from './disabled';
 import Focusable from './focusable';
 import Hoverable from './hoverable';
+import HtmlFor from './htmlFor';
 import Infos from './infos';
 import Required from './required';
 import Status from './status';
@@ -14,6 +15,7 @@ import BorderSource from './border?raw';
 import DisabledSource from './disabled?raw';
 import FocusableSource from './focusable?raw';
 import HoverableSource from './hoverable?raw';
+import HtmlForSource from './htmlFor?raw';
 import InfosSource from './infos?raw';
 import RequiredSource from './required?raw';
 import StatusSource from './status?raw';
@@ -25,6 +27,7 @@ export const Sources = {
     Disabled: DisabledSource,
     Focusable: FocusableSource,
     Hoverable: HoverableSource,
+    HtmlFor: HtmlForSource,
     Infos: InfosSource,
     Required: RequiredSource,
     Status: StatusSource
@@ -36,6 +39,7 @@ export default {
     Disabled,
     Focusable,
     Hoverable,
+    HtmlFor,
     Infos,
     Required,
     Status